Tag: V00-07-07 2010-08-12 Andy Salnikov - O2OXtcMerger now changes datagram clock time - l1OffsetSec is now double so can have sub-second precision - o2o-scanner has --l1-offset option too
git-svn-id: https://pswww.slac.stanford.edu/svn-readonly/psdmrepo@1167 b967ad99-d558-0410-b138-e0f6c56caec7

+++ O2OTranslator/trunk/app/o2o-translate.cpp
@@ -93,7 +93,7 @@ private:
   AppCmdOpt<std::string>      m_experiment ;
   AppCmdOptBool               m_extGroups ;
   AppCmdOpt<std::string>      m_instrument ;
-  AppCmdOpt<int>              m_l1offset ;
+  AppCmdOpt<double>           m_l1offset ;
   AppCmdOpt<std::string>      m_mdConnStr ;
   AppCmdOptNamedValue<O2OXtcMerger::MergeMode> m_mergeMode ;
   AppCmdOptList<std::string>  m_metadata ;
